Converted the CLI project into the SymbioticTS.MSBuild package
While the SymbioticTS.Build package, which includes an MSBuild task, is great, it falls apart due to MSBuild node reuse, which was enabled by default at some point in the last few years. This keeps the process alive; including the file handles on assemblies that were loaded, which prevents us from being able to build again due to locked files. There is currently no way to unload assemblies once loaded. This is supposed to be supported in .NET Core 3.0, but we'd also need it in a version of netstandard supported by MSBuild.

As a workaround, I am repurposing the CLI which will run using the `Exec` task and terminate releasing all handles.

The new SymbioticTS.MSBuild package will now be the preferred method for hooking up MSBuild integration until the SymbioticTS.Build package can be supported correctly.